+++ frontend/test/frontend_tests/runner.cljs
@@ -16,6 +16,7 @@
    [frontend-tests.tokens.logic.token-actions-test]
    [frontend-tests.tokens.logic.token-data-test]
    [frontend-tests.tokens.style-dictionary-test]
+   [frontend-tests.tokens.workspace-tokens-remap-test]
    [frontend-tests.util-object-test]
    [frontend-tests.util-range-tree-test]
    [frontend-tests.util-simple-math-test]
@@ -49,4 +50,5 @@
    'frontend-tests.util-object-test
    'frontend-tests.util-range-tree-test
    'frontend-tests.util-simple-math-test
+   'frontend-tests.tokens.workspace-tokens-remap-test
    'frontend-tests.worker-snap-test))